feat: Add pagination to getAllTrx endpoint
The code changes in `transactionsController.ts` and `transactionService.ts` introduce pagination functionality to the `getAllTrx` endpoint. The `limit` and `skip` parameters are now accepted from the request query and used to limit the number of transactions returned and skip a certain number of transactions, respectively. This enhancement allows for more efficient retrieval of transactions and improves the user experience.