feat: Improve getAllAccounts performance with aggregation pipeline
The code changes in `index.ts` of the `accountService` directory modify the `getAllAccounts` method to use the aggregation pipeline instead of the `find` method. This change includes a `$sort` stage to sort the accounts by `createdAt` in descending order. By using the aggregation pipeline, the performance of retrieving all accounts is improved, resulting in a better user experience.