Refactor build.js to enable minification and disable sourcemap
The code changes in build.js modify the build configuration to enable minification and disable sourcemap generation. This improves the performance of the build process by reducing the size of the output file and removing unnecessary sourcemap generation. The purpose of these changes is to optimize the production build and improve overall application performance.